requirements.txt
flask
azure-ai-documentintelligence
werkzeug